Show test status dynamically

+++ YoYo.Domain/Entities/FitnessRating.cs
@@ -16,7 +16,7 @@ namespace YoYo.Domain.Entities
         public int SpeedLevel { get; set; }
         public int ShuttleNo { get; set; }
         public decimal Speed { get; set; }
-        public decimal LevelTime { get; set; }
+        public int LevelTime { get; set; }
         public string CommulativeTime { get; set; }
         public string StartTime { get; set; }
         public decimal ApproxVo2Max { get; set; }
